+++ app/src/main/AndroidManifest.xml
@@ -31,7 +31,7 @@
         </provider>
 
         <activity
-            android:name=".MainActivity"
+            android:name=".ui.MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.NatureAlbum">
