+++ app/src/main/java/com/and04/naturealbum/ui/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.and04.naturealbum
+package com.and04.naturealbum.ui
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
