+++ app/src/main/java/com/and04/naturealbum/ui/NatureAlbumApp.kt
@@ -1,4 +1,4 @@
-package com.and04.naturealbum
+package com.and04.naturealbum.ui
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.platform.LocalContext
@@ -6,6 +6,7 @@ import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
+import com.and04.naturealbum.ui.home.HomeScreen
 import com.and04.naturealbum.ui.savephoto.SavePhotoScreen
 import com.and04.naturealbum.ui.theme.NatureAlbumTheme
 import java.io.File
