+++ app/src/main/java/com/and04/naturealbum/ui/component/MyDialog.kt
@@ -1,4 +1,4 @@
-package com.and04.naturealbum
+package com.and04.naturealbum.ui.component
 
 import androidx.annotation.DrawableRes
 import androidx.annotation.StringRes
@@ -7,9 +7,9 @@ import androidx.compose.material3.Icon
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
+import com.and04.naturealbum.R
 
 
 @Composable
